training/requirements.txt
torch==2.8.0
transformers==4.46.3
datasets==3.1.0
accelerate==1.1.1
scikit-learn==1.5.2
pandas==2.2.3
numpy==1.26.4
google-cloud-aiplatform==1.71.1
google-cloud-storage==2.18.2
gcsfs==2024.9.0
